systemd: default to use libsystemd if build with support for it
We may assume that if an user does build conntrackd with such feature, is with
the intention to use it. So, if that's the case, default to use it.

This eases some downstream use cases when dealing with default configs to
be shipped to final users.

This could be a mid-point solution, given some users are asking for a full
revert of commit c01d0d9138112ec95ee316385ea2687dd94fa4e3.

Signed-off-by: Arturo Borrero Gonzalez <arturo@netfilter.org>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>