conntrack: default to unspec family for dualstack setups
2bcbae4c14b2 ("conntrack: -f family filter does not work") restored the
fallback to IPv4 if -f is not specified, which was the original
behaviour.

This patch modifies the default to use the unspec family if -f is not
specified for the following ct commands:

- list
- update
- delete
- get

(these two commands below do not support for -f though, but in case this is
extended in the future to support it):

- flush
- event

The existing code that parses IPv4 and IPv6 addresses already infers the
family, which simplifies the introduction of this update.

The expect commands are not updated, they still require many mandatory
options for filtering.

This patch includes a few test updates too.

Based on patch from Mikhail Sennikovsky.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>