conntrack: add struct ct_tmpl
Remove the global template object, add it to struct ct_cmd. This patch
prepares for the batch support.

The global cur_tmpl pointer is used to access the template from the
callbacks and the exit_error() path.

Note that it should be possible to remove this global cur_tmpl pointer
by passing the new command object as parameter to the callbacks and
exit_error().

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>