tests: introduce new python-based framework for running tests
This test suite should help us develop better tests for conntrack-tools in general and conntrackd
in particular.

The framework is composed of a runner script, written in python3, and 3 yaml files for
configuration and testcase definition:

 - scenarios.yaml: contains information on network scenarios for tests to use
 - tests.yaml: contains testcase definition
 - env.yaml: contains default values for environment variables

The test cases can be anything, from a simple command to an external script call to perform more
complex operations. See follow-up patches to know more on how this works.

The plan is to replace or call from this framework the other testsuites in this tree.

The runner script is rather simple, and it should be more or less straight forward to use it.
On Debian machines, it requires the *python3-yaml* package to be installed as a dependency.

For reference, here are the script options:

=== 8< ===
$ tests/conntrackd/conntrackd-tests.py --help
usage: conntrackd-tests.py [-h] [--tests-file TESTS_FILE]
				[--scenarios-file SCENARIOS_FILE]
				[--env-file ENV_FILE]
				[--single SINGLE]
				[--start-scenario START_SCENARIO]
				[--stop-scenario STOP_SCENARIO]
				[--debug]

Utility to run tests for conntrack-tools

optional arguments:
  -h, --help            show this help message and exit
  --tests-file TESTS_FILE
                        File with testcase definitions. Defaults to 'tests.yaml'
  --scenarios-file SCENARIOS_FILE
                        File with configuration scenarios for tests. Defaults to 'scenarios.yaml'
  --env-file ENV_FILE   File with environment variables for scenarios/tests. Defaults to 'env.yaml'
  --single SINGLE       Execute a single testcase and exit. Use this for developing testcases
  --start-scenario START_SCENARIO
                        Execute scenario start commands and exit. Use this for developing testcases
  --stop-scenario STOP_SCENARIO
                        Execute scenario stop commands and exit. Use this for cleanup
  --debug               debug mode
=== 8< ===

To run it, simply use:

=== 8< ===
$ cd tests/conntrackd/ ; sudo ./conntrackd-tests.py
[..]
=== 8< ===

Signed-off-by: Arturo Borrero Gonzalez <arturo@netfilter.org>