conntrack: add support for chaintoolong stat counter
While at it, also fix up commit
5ec55c2a0a264 ("conntrack: add support for CLASH_RESOLVED counter").

The intention was to also print unknown statistic counters provided
the attributes are of type u32, but mnl_attr_type_valid() needs the
correct max-attr number for this to work.

With this fix in place, conntrack -S shows:

cpu=3    found=0 inv[..] clash_resolve=0 unknown1=8200

on a modified kernel.

Signed-off-by: Florian Westphal <fw@strlen.de>