build: only require bison and flex if the generated files do not exist
automake recommends including the files generated by bison and flex in
distribution tar-balls and runs bison and flex during `make dist` to
generate them.  Thus, in the normal case where the software is being
compiled by an end-user, the generated files already exist and bison and
flex are not required.  Therefore, amend the configure script only to
require them if the generated files do not exist.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>